# docs/OPENAPI.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Dossier Public API",
    "version": "1.0.0-draft",
    "description": "Public API contract for Dossier. This file is a pre-build review artifact and does not include application code. Scope: public endpoints only. Internal admin endpoints such as /internal/registry/import and /internal/registry/verify are intentionally omitted. OpenAPI path parameters use {destinationId} as the standard translation of the product spec notation :destinationId."
  },
  "servers": [
    {
      "url": "https://api.dossier.example"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/v1/ai/transcribe": {
      "post": {
        "summary": "Create a transcript from captured audio",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranscribeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transcript created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscribeResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "413": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "422": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/v1/ai/extract": {
      "post": {
        "summary": "Extract facts, people, places, business matches, and timeline details",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Fact set created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "422": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/v1/ai/draft": {
      "post": {
        "summary": "Create a route-aware draft report",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Draft created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "422": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "ErrorResponse": {
        "description": "Error response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "ok": false,
              "request_id": "req_01JQ9YZ0QMC1PR3Q7PSQMX9GSQ",
              "error": {
                "code": "route_data_unavailable",
                "message": "Route data could not be loaded right now.",
                "retryable": true,
                "details": {
                  "endpoint": "/v1/routes/recommend"
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
      "SuccessBase": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "ok",
          "request_id"
        ],
        "properties": {
          "ok": {
            "const": true
          },
          "request_id": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "ok",
          "request_id",
          "error"
        ],
        "properties": {
          "ok": {
            "const": false
          },
          "request_id": {
            "type": "string"
          },
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "code",
              "message",
              "retryable",
              "details"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "retryable": {
                "type": "boolean"
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      },
...
      "InlineAudioUpload": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "upload_mode",
          "filename",
          "mime_type",
          "size_bytes",
          "content_base64"
        ],
        "properties": {
          "upload_mode": {
            "const": "inline_base64"
          },
          "filename": {
            "type": "string"
          },
          "mime_type": {
            "type": "string"
          },
          "size_bytes": {
            "type": "integer",
            "minimum": 1
          },
          "content_base64": {
            "type": "string"
          }
        }
      },
      "ObjectStorageAudioUpload": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "upload_mode",
          "filename",
          "mime_type",
          "size_bytes",
          "object_key"
        ],
        "properties": {
          "upload_mode": {
            "const": "object_storage_reference"
          },
          "filename": {
            "type": "string"
          },
          "mime_type": {
            "type": "string"
          },
          "size_bytes": {
            "type": "integer",
            "minimum": 1
          },
          "object_key": {
            "type": "string"
          }
        }
      },
      "TranscriptSegment": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "start_ms",
          "end_ms",
          "speaker_label",
          "text",
          "confidence"
        ],
        "properties": {
          "start_ms": {
            "type": "integer",
            "minimum": 0
          },
          "end_ms": {
            "type": "integer",
            "minimum": 0
          },
          "speaker_label": {
            "type": [
              "string",
              "null"
            ]
          },
          "text": {
            "type": "string"
          },
          "confidence": {
            "type": [
              "number",
              "null"
            ],
            "minimum": 0,
            "maximum": 1
          }
        }
      },
      "TranscriptDocument": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "full_text",
          "language",
          "segment_count",
          "segments"
        ],
        "properties": {
          "full_text": {
            "type": "string"
          },
          "language": {
            "type": [
              "string",
              "null"
            ]
          },
          "segment_count": {
            "type": "integer",
            "minimum": 0
          },
          "segments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TranscriptSegment"
            }
          }
        }
      },
      "ModelMetadata": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "provider",
          "model",
          "purpose",
          "requested_at",
          "completed_at",
          "latency_ms"
        ],
        "properties": {
          "provider": {
            "const": "openai"
          },
          "model": {
            "type": "string"
          },
          "purpose": {
            "type": "string",
            "enum": [
              "transcribe",
              "extract",
              "draft",
              "route_reasoning"
            ]
          },
          "requested_at": {
            "type": "string",
            "format": "date-time"
          },
          "completed_at": {
            "type": "string",
            "format": "date-time"
          },
          "latency_ms": {
            "type": "integer",
            "minimum": 0
          },
          "input_audio_seconds": {
            "type": [
              "number",
              "null"
            ],
            "minimum": 0
          },
          "input_characters": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          },
          "input_tokens": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          },
          "output_tokens": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          }
        }
      },
      "TranscribeRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "incident_id",
          "source_evidence_id",
          "source_evidence_sha256",
          "upload"
        ],
        "properties": {
          "incident_id": {
            "type": "string",
            "format": "uuid"
          },
          "source_evidence_id": {
            "type": "string",
            "format": "uuid"
          },
          "source_evidence_sha256": {
            "type": "string"
          },
          "upload": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/InlineAudioUpload"
              },
              {
                "$ref": "#/components/schemas/ObjectStorageAudioUpload"
              }
            ]
          },
          "language_hint": {
            "type": [
              "string",
              "null"
            ]
          },
          "include_timestamps": {
            "type": "boolean",
            "default": true
          }
        }
      },
      "TranscribeResponseData": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "incident_id",
          "source_evidence_id",
          "transcript_evidence_id",
          "transcript",
          "model_metadata",
          "warnings"
        ],
        "properties": {
          "incident_id": {
            "type": "string",
            "format": "uuid"
          },
          "source_evidence_id": {
            "type": "string",
            "format": "uuid"
          },
          "transcript_evidence_id": {
            "type": "string",
            "format": "uuid"
          },
          "transcript": {
            "$ref": "#/components/schemas/TranscriptDocument"
          },
          "model_metadata": {
            "$ref": "#/components/schemas/ModelMetadata"
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "TranscribeResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SuccessBase"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "$ref": "#/components/schemas/TranscribeResponseData"
              }
            }
          }
        ]
      },
      "FactSet": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "fact_set_id",
          "incident_type",
          "people",
          "places",
          "businesses",
          "phones",
          "dates",
          "amounts",
          "timeline",
          "key_facts",
          "reviewed_by_user"
        ],
        "properties": {
          "fact_set_id": {
            "type": "string",
            "format": "uuid"
          },
          "incident_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "people": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "places": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "businesses": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "phones": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "dates": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "amounts": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "timeline": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": false,
              "required": [
                "time_label",
                "description"
              ],
              "properties": {
                "time_label": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                }
              }
            }
          },
          "key_facts": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "reviewed_by_user": {
            "type": "boolean"
          }
        }
      },
      "ExtractContext": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "location_address",
          "confirmed_place_id",
          "confirmed_place_name",
          "confirmed_place_phone"
        ],
        "properties": {
          "location_address": {
            "type": [
              "string",
              "null"
            ]
          },
          "confirmed_place_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "confirmed_place_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "confirmed_place_phone": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "ExtractRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "incident_id",
          "transcript_evidence_id",
          "transcript",
          "context"
        ],
        "properties": {
          "incident_id": {
            "type": "string",
            "format": "uuid"
          },
          "transcript_evidence_id": {
            "type": "string",
            "format": "uuid"
          },
          "transcript": {
            "$ref": "#/components/schemas/TranscriptDocument"
          },
          "context": {
            "$ref": "#/components/schemas/ExtractContext"
          }
        }
      },
      "ExtractResponseData": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "incident_id",
          "transcript_evidence_id",
          "fact_set",
          "model_metadata",
          "warnings"
        ],
        "properties": {
          "incident_id": {
            "type": "string",
            "format": "uuid"
          },
          "transcript_evidence_id": {
            "type": "string",
            "format": "uuid"
          },
          "fact_set": {
            "$ref": "#/components/schemas/FactSet"
          },
          "model_metadata": {
            "$ref": "#/components/schemas/ModelMetadata"
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ExtractResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SuccessBase"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "$ref": "#/components/schemas/ExtractResponseData"
              }
            }
          }
        ]
      },
      "MailingAddress": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "line1",
          "line2",
          "city",
          "state",
          "postal_code",
          "country"
        ],
        "properties": {
          "line1": {
            "type": "string"
          },
          "line2": {
            "type": [
              "string",
              "null"
            ]
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "postal_code": {
            "type": "string"
          },
          "country": {
            "const": "US"
          }
        }
      },
      "Destination": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "destination_id",
          "destination_name",
          "destination_type",
          "jurisdiction",
          "categories_handled",
          "intake_methods",
          "complaint_url",
          "email",
          "phone",
          "mailing_address",
          "source_url",
          "last_verified_date",
          "trust_level",
          "notes_required_fields",
          "notes_required_documents"
        ],
        "properties": {
          "destination_id": {
            "type": "string",
            "format": "uuid"
          },
          "destination_name": {
            "type": "string"
          },
          "destination_type": {
            "type": "string",
            "enum": [
              "business",
              "corporate",
              "local_agency",
              "state_agency",
              "federal_agency",
              "regulator",
              "law_enforcement",
              "consumer_org",
              "other"
            ]
          },
          "jurisdiction": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "country",
              "state",
              "county",
              "city"
            ],
            "properties": {
              "country": {
                "const": "US"
              },
              "state": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "county": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "city": {
                "type": [
                  "string",
                  "null"
                ]
              }
            }
          },
          "categories_handled": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "intake_methods": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "web_form",
                "email",
                "phone",
                "mail",
                "portal",
                "in_person"
              ]
            }
          },
          "complaint_url": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "email": {
            "type": [
              "string",
              "null"
            ],
            "format": "email"
          },
          "phone": {
            "type": [
              "string",
              "null"
            ]
          },
          "mailing_address": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/MailingAddress"
              },
              {
                "type": "null"
              }
            ]
          },
          "source_url": {
            "type": "string",
            "format": "uri"
          },
          "last_verified_date": {
            "type": "string",
            "format": "date"
          },
          "trust_level": {
            "type": "string",
            "enum": [
              "official",
              "verified",
              "directory",
              "unconfirmed"
            ]
          },
          "notes_required_fields": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "notes_required_documents": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "RouteRecommendation": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "destination_id",
          "destination_name_snapshot",
          "destination_type_snapshot",
          "route_category",
          "rank",
          "reason",
          "source_label",
          "source_url",
          "last_verified_date",
          "trust_level",
          "intake_methods_snapshot",
          "required_documents_snapshot",
          "available_actions",
          "destination"
        ],
        "properties": {
          "destination_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "destination_name_snapshot": {
            "type": "string"
          },
          "destination_type_snapshot": {
            "type": "string"
          },
          "route_category": {
            "type": "string",
            "enum": [
              "Business",
              "Local",
              "State",
              "Federal",
              "Other verified routes"
            ]
          },
          "rank": {
            "type": "integer",
            "minimum": 1
          },
          "reason": {
            "type": "string"
          },
          "source_label": {
            "type": "string"
          },
          "source_url": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "last_verified_date": {
            "type": [
              "string",
              "null"
            ],
            "format": "date"
          },
          "trust_level": {
            "type": "string",
            "enum": [
              "official",
              "verified",
              "directory",
              "unconfirmed"
            ]
          },
          "intake_methods_snapshot": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "web_form",
                "email",
                "phone",
                "mail",
                "portal",
                "in_person"
              ]
            }
          },
          "required_documents_snapshot": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "available_actions": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "open_form",
                "call",
                "email",
                "share_packet",
                "export_packet",
                "save_for_later"
              ]
            }
          },
          "destination": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Destination"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
...
      "AttachmentReference": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "evidence_id",
          "label",
          "mime_type",
          "sha256"
        ],
        "properties": {
          "evidence_id": {
            "type": "string",
            "format": "uuid"
          },
          "label": {
            "type": "string"
          },
          "mime_type": {
            "type": "string"
          },
          "sha256": {
            "type": "string"
          }
        }
      },
      "DraftPacket": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "draft_packet_id",
          "incident_id",
          "destination_id",
          "subject",
          "body",
          "attachments",
          "version",
          "approved"
        ],
        "properties": {
          "draft_packet_id": {
            "type": "string",
            "format": "uuid"
          },
          "incident_id": {
            "type": "string",
            "format": "uuid"
          },
          "destination_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "subject": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AttachmentReference"
            }
          },
          "version": {
            "type": "integer",
            "minimum": 1
          },
          "approved": {
            "type": "boolean"
          }
        }
      },
      "DraftRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "incident_id",
          "fact_set",
          "selected_route"
        ],
        "properties": {
          "incident_id": {
            "type": "string",
            "format": "uuid"
          },
          "fact_set": {
            "$ref": "#/components/schemas/FactSet"
          },
          "selected_route": {
            "$ref": "#/components/schemas/RouteRecommendation"
          },
          "transcript_excerpt": {
            "type": [
              "string",
              "null"
            ]
          },
          "desired_tone": {
            "type": "string",
            "default": "plain_serious"
          }
        }
      },
      "DraftResponseData": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "incident_id",
          "selected_route",
          "draft_packet",
          "model_metadata",
          "warnings"
        ],
        "properties": {
          "incident_id": {
            "type": "string",
            "format": "uuid"
          },
          "selected_route": {
            "$ref": "#/components/schemas/RouteRecommendation"
          },
          "draft_packet": {
            "$ref": "#/components/schemas/DraftPacket"
          },
          "model_metadata": {
            "$ref": "#/components/schemas/ModelMetadata"
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "DraftResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SuccessBase"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "$ref": "#/components/schemas/DraftResponseData"
              }
            }
          }
        ]
      },
...
    }
  }
}